fix: v5.0 (#260)
